+++ lib/widgets/fields/build_model_textfield.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+
+class BuildModelTextField extends StatelessWidget {
+  final String label;
+  final String value;
+  final ValueChanged<String> onChanged;
+
+  const BuildModelTextField({
+    required this.label,
+    required this.value,
+    required this.onChanged,
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(8.0),
+      child: TextField(
+        decoration: InputDecoration(
+          labelText: label,
+          border: const OutlineInputBorder(),
+          isDense: true,
+        ),
+        controller: TextEditingController(text: value),
+        onChanged: onChanged,
+      ),
+    );
+  }
+}